updated the conflict occuring codes
